03_Fund_Price_AEGON_import.R: migrating to BAMOSZ fund price acquisition

+++ 01_scheduled_jobs/03_Fund_Price_AEGON_import.R
@@ -2,7 +2,8 @@ message("----------------------------------------")
 message(paste("Job starts:",Sys.time()))
 
 message("Importing necessary packages..")
-library(gdata)
+library(RCurl)
+library(XML)
 
 message("Importing functions..")
 source("/srv/shiny-server/finapp/f.R")
@@ -11,51 +12,53 @@ message("Other setups..")
 options(scipen = 999, digits = 4, encoding = "iso-8859-2")
 
 message("Creating import control table..")
-fp_import_ctrltbl <- psqlQuery("SELECT f.id fund_id, f.fund_name, f.source_id, COALESCE(fpv.value_date,f.inception_date-1)+1 import_date_from
+fp_import_ctrltbl <- psqlQuery("SELECT f.id fund_id, f.isin, f.fund_name, f.source_id, COALESCE(fpv.value_date,f.inception_date-1)+1 import_date_from
                                FROM app.fund f
                                LEFT OUTER JOIN app.fund_price_recent_vw fpv ON f.id=fpv.fund_id
-                               WHERE f.description like 'AEGON%'")$result
+                               WHERE f.ISIN like 'HU%'")$result
 
 message("Content of control table:")
 print(fp_import_ctrltbl)
 
 message("Importing..")
-price_tbl_pre <- data.frame(date=character(),fund_id=character(),fund_name=character(),price=character(),stringsAsFactors=F)
-for(i in 1:nrow(fp_import_ctrltbl)){
-    message(paste("processing: ", fp_import_ctrltbl[i,2]))
-    message(paste("downloading..",
-                sprintf("https://www.aegonalapkezelo.hu/elemzes/grafikonrajzolo/?id%%5B%%5D=%s&mode=download&min_date=%s&max_date=%s",
-                        fp_import_ctrltbl[i,3],
-                        fp_import_ctrltbl[i,4],
-                        Sys.Date())))
-    download.file(url=sprintf("https://www.aegonalapkezelo.hu/elemzes/grafikonrajzolo/?id%%5B%%5D=%s&mode=download&min_date=%s&max_date=%s",
-                              fp_import_ctrltbl[i,3],
-                              fp_import_ctrltbl[i,4],
-                              Sys.Date()),
-                  destfile = "tmp.xls",
-                  method = "curl")
-    if(length(readLines(xls2csv("tmp.xls")))==0){
-        price_tbl_pre <- rbind(price_tbl_pre,
-                               data.frame(date=character(),fund_id=character(),fund_name=character(),price=character(),stringsAsFactors=F))
-    }
-    else {
-        fileContent <- read.xls("tmp.xls",sheet = 1,stringsAsFactors=F)
-        price_tbl_pre <- rbind(price_tbl_pre,
-                               cbind(fileContent[-1,1],fp_import_ctrltbl[i,1],fp_import_ctrltbl[i,2],fileContent[-1,3]))
-    }
-    file.remove("tmp.xls")
-}
+for(i_fundid in fp_import_ctrltbl$fund_id){
+    message(paste0("..processing: ",fp_import_ctrltbl[fp_import_ctrltbl$fund_id==i_fundid,"fund_name"]))
+    retrieve_url <- sprintf("http://www.bamosz.hu/en/alapoldal?isin=%s",fp_import_ctrltbl[fp_import_ctrltbl$fund_id==i_fundid,"isin"])
+    message(paste0("..reading in from URL: ",retrieve_url))
+    web_import <- getURL(retrieve_url)
+    df_import <- readHTMLTable(web_import, trim = T, header = F, as.data.frame = T, stringsAsFactors=F)[[7]][-1,c(1:4)]
+    
+    colnames(df_import) <- c("date","price","net_asset_value","paid_dividend")
+    df_import$date <- as.Date(df_import$date,"%Y.%m.%d.")
+    df_import$price <- as.numeric(df_import$price)
+    df_import$net_asset_value <- as.numeric(gsub(",","",df_import$net_asset_value))
+    df_import$paid_dividend <- as.numeric(df_import$paid_dividend)
+    df_import$fund_id <- i_fundid
+    df_import$fund_name <- fp_import_ctrltbl[fp_import_ctrltbl$fund_id==i_fundid,"fund_name"]
 
-message("Downloaded data to be imported:")
-print(price_tbl_pre)
-
-truncateStatus <- psqlQuery("TRUNCATE TABLE app.ld_fund_price")
-message(paste("Truncate app.ld_fund_price..",truncateStatus$errorMsg))
-
-ldInsertStatus <- psqlInsert(price_tbl_pre, "ld_fund_price")
-message(paste("Insert into app.ld_fund_price..",ldInsertStatus$errorMsg))
-
-finalInsertStatus <- psqlQuery("INSERT INTO app.fund_price (fund_id,value_date,price)
+    message("Fetched data (first 7): ")
+    print(head(df_import,7))
+    
+    db_cont <- psqlQuery(sprintf("SELECT value_date FROM app.fund_price WHERE fund_id=%i",i_fundid))$result
+    
+    df_to_load <- df_import[!(df_import$date %in% db_cont$value_date),]
+    df_to_load$net_asset_value <- NULL
+    df_to_load$paid_dividend <- NULL
+    df_to_load <- data.frame(date=df_to_load$date,
+                             fund_id=df_to_load$fund_id,
+                             fund_name=df_to_load$fund_name,
+                             price=df_to_load$price)
+    
+    message("Content to load in db: ")
+    print(df_to_load)
+    
+    if(nrow(df_to_load)>0)
+    {
+        truncateStatus <- psqlQuery("TRUNCATE TABLE app.ld_fund_price;")
+        message(paste0("Truncate app.ld_fund_price table........",truncateStatus$errorMsg))
+        insertLdStatus <- psqlInsert(df_to_load,"ld_fund_price")
+        message(paste0("Insert into app.ld_fund_price table........",insertLdStatus$errorMsg))
+        finalInsertStatus <- psqlQuery("INSERT INTO app.fund_price (fund_id,value_date,price)
                                SELECT
                                ldfp.fund_id::int
                                ,to_date(ldfp.date, 'yyyy-mm-dd')
@@ -64,10 +67,9 @@ finalInsertStatus <- psqlQuery("INSERT INTO app.fund_price (fund_id,value_date,p
                                LEFT OUTER JOIN app.fund_price fp ON ldfp.fund_id::int=fp.fund_id
                                AND to_date(ldfp.date, 'yyyy-mm-dd')=fp.value_date
                                WHERE fp.id IS NULL")
-
-message(paste("Insert into app.fund_price..",finalInsertStatus$errorMsg))
-
-updPortfRefDate <- psqlQuery("
+        
+        message(paste("Insert into app.fund_price..",finalInsertStatus$errorMsg))
+        updPortfRefDate <- psqlQuery("
                             UPDATE app.portfolio SET last_valuation_date=u.vd
                             FROM (
                                                          SELECT 
@@ -86,8 +88,10 @@ updPortfRefDate <- psqlQuery("
                             ) AS u
                                                          WHERE app.portfolio.id = u.portfolio_id                             
                              ")
-
-message(paste("Update app.portfolio with new portfolio valuation date..",updPortfRefDate$errorMsg))
+        message(paste("Update app.portfolio with new portfolio valuation date..",updPortfRefDate$errorMsg))
+    }
+    message("\n")
+}
 
 refreshMatVw1 <- psqlQuery("REFRESH MATERIALIZED VIEW app.portfolio_return_calc_mvw")
 refreshMatVw2 <- psqlQuery("REFRESH MATERIALIZED VIEW app.portf_acc_return_calc_mvw")
